Separate source generator infrastructure
