+++ UnitsNet.Modular/UnitsNet.Modular/AuthoringContracts.cs
