+++ UnitsNet.Modular/UnitsNet.Modular/QuantityInfo.cs
@@ -4,6 +4,7 @@ using System.Collections.Frozen;
 using System.ComponentModel;
 using System.Diagnostics.CodeAnalysis;
 using UnitsNet.Modular;
+using UnitsNet.Modular.SourceGen;
 
 namespace UnitsNet;
